pyproject.toml
[project]
name = "meeting-mood-tracker"
version = "0.1.0"
description = "Meeting Mood Tracker API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "pytest",
    "httpx",
    "python-dotenv>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "harness"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"] # pycodestyle, pyflakes, isort
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
